Remove unneccessary message padding now that MC-40 is fixed.
A Minecraft bug existed that caused incorrect offsets for
center-justified text that included color codes. This occured on signs,
player nametags, and kick message text. It was fixed in 1.4.5 and
colored text is no longer offset, and as a result the whitespace added
to compensate for the bug is no longer required.

See MC-40: https://mojang.atlassian.net/browse/MC-40